feat: Add UserEducationInfoCard component
Added the UserEducationInfoCard component to display education information of a user. The component takes in an array of educationInfo objects and renders them in a structured format. This component is used in the Profile page to display the user's education history. The education information includes fields such as field of study, institute name, start and end dates, and a description. The component has been added to the Profile page to showcase the user's education along with other profile information.