feat: Add modal for applying to projects
Added a dialog modal to allow users to write and submit a cover letter when applying to projects. The modal includes a form with validation for the message length and Save/Cancel buttons for submission.

Added application page to the user profile so users can view and mange their applications